feat: body layout refactoring and adding mici layout
