+++ selfdrive/ui/body/layouts/__init__.py
@@ -1,0 +1,3 @@
+from openpilot.selfdrive.ui.body.layouts.body import BodyLayout
+from openpilot.selfdrive.ui.body.layouts.body_sidebar import BodySidebar, BODY_SIDEBAR_HEIGHT
+from openpilot.selfdrive.ui.body.layouts.home import MiciBodyHomeLayout
