+++ selfdrive/ui/layouts/main.py
@@ -3,14 +3,13 @@ from enum import IntEnum
 import cereal.messaging as messaging
 from openpilot.system.ui.lib.application import gui_app
 from openpilot.selfdrive.ui.layouts.sidebar import Sidebar, SIDEBAR_WIDTH
-from openpilot.selfdrive.ui.layouts.body.body import BodyLayout
-from openpilot.selfdrive.ui.layouts.body.body_sidebar import BodySidebar, BODY_SIDEBAR_HEIGHT
 from openpilot.selfdrive.ui.layouts.home import HomeLayout
 from openpilot.selfdrive.ui.layouts.settings.settings import SettingsLayout, PanelType
 from openpilot.selfdrive.ui.onroad.augmented_road_view import AugmentedRoadView
 from openpilot.selfdrive.ui.ui_state import device, ui_state
 from openpilot.system.ui.widgets import Widget
 from openpilot.selfdrive.ui.layouts.onboarding import OnboardingWindow
+from openpilot.selfdrive.ui.body.layouts import BodyLayout, BodySidebar, BODY_SIDEBAR_HEIGHT
 
 
 class MainState(IntEnum):
